remove index.ts files, make constants class
