+++ src/app/app.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
-import { AppController } from './controllers';
-import { MyAccountModule } from '../my-account';
-import { UsersModule } from '../users';
-import { AuthModule } from '../auth';
-import { DbModule } from '../db';
+import { AppController } from './controllers/app.controller';
+import { MyAccountModule } from '../my-account/my-account.module';
+import { UsersModule } from '../users/users.module';
+import { AuthModule } from '../auth/auth.module';
+import { DbModule } from '../db/db.module';
 import { ConfigModule } from '@nestjs/config';
 import { validate } from '../../env.validation';
