--- src/app/controllers/index.ts
@@ -1,1 +1,0 @@
-export { AppController } from './app.controller';
