+++ src/auth/auth.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
-import { AuthController } from './controllers';
-import { AuthService } from './services';
+import { AuthController } from './controllers/auth.controller';
+import { AuthService } from './services/auth.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Authentication } from './entities';
-import { UsersModule } from '../users';
+import { Authentication } from './entities/auth.entity';
+import { UsersModule } from '../users/users.module';
 
 @Module({
   imports: [UsersModule, TypeOrmModule.forFeature([Authentication])],
