+++ src/auth/controllers/auth.controller.ts
@@ -1,9 +1,10 @@
 import { Controller, Post, Body, HttpCode, HttpStatus } from '@nestjs/common';
 import { ApiOperation, ApiTags, ApiOkResponse, ApiInternalServerErrorResponse } from '@nestjs/swagger';
-import { AuthService } from '../services';
-import { CreateAuthDTO } from '../dtos';
-import { User } from '../../users/entities';
-import { CreateUserDTO, UserDTO } from '../../users/dtos';
+import { AuthService } from '../services/auth.service';
+import { CreateAuthDTO } from '../dtos/create-auth.dto';
+import { User } from '../../users/entities/user.entity';
+import { UserDTO } from '../../users/dtos/user.dto';
+import { CreateUserDTO } from '../../users/dtos/create-user.dto';
 
 @Controller('auth')
 export class AuthController {
