--- src/auth/controllers/index.ts
@@ -1,1 +1,0 @@
-export { AuthController } from './auth.controller';
