--- src/auth/dtos/index.ts
@@ -1,2 +1,0 @@
-export { AuthDTO } from './auth.dto';
-export { CreateAuthDTO } from './create-auth.dto';
