--- src/auth/entities/index.ts
@@ -1,1 +1,0 @@
-export { Authentication } from './auth.entity';
