--- src/auth/index.ts
@@ -1,1 +1,0 @@
-export { AuthModule } from './auth.module';
