--- src/auth/services/index.ts
@@ -1,1 +1,0 @@
-export { AuthService } from './auth.service';
