--- src/auth/subscribers/index.ts
@@ -1,1 +1,0 @@
-export { AuthSubscriber } from './auth-subscriber';
