--- src/db/index.ts
@@ -1,1 +1,0 @@
-export { DbModule } from './db.module';
