+++ src/main.ts
@@ -1,8 +1,8 @@
 import 'reflect-metadata';
 import { NestFactory } from '@nestjs/core';
 import { SwaggerModule } from '@nestjs/swagger';
-import { swaggerConfig } from './utils/swagger-config';
-import { AppModule } from './app';
+import { swaggerConfig } from './utils/configs/swagger-config';
+import { AppModule } from './app/app.module';
 import { ValidationPipe, ClassSerializerInterceptor } from '@nestjs/common';
 import { Reflector } from '@nestjs/core';
 import { ConfigService } from '@nestjs/config';
