--- src/my-account/controllers/index.ts
@@ -1,1 +1,0 @@
-export { MyAccountController } from "./me.controller";
