--- src/my-account/index.ts
@@ -1,1 +1,0 @@
-export { MyAccountModule } from "./me.module";
