--- src/my-account/services/index.ts
@@ -1,1 +1,0 @@
-export { MyAccountService } from "./me.service";
