+++ src/users/dtos/create-user.dto.ts
@@ -1,5 +1,5 @@
 import { IsString, IsDateString, IsNotEmpty } from 'class-validator';
-import { CreateAuthDTO } from '../../auth/dtos';
+import { CreateAuthDTO } from '../../auth/dtos/create-auth.dto';
 import { ApiProperty } from '@nestjs/swagger';
 
 export class CreateUserDTO extends CreateAuthDTO {
