--- src/users/dtos/index.ts
@@ -1,2 +1,0 @@
-export { UserDTO } from './user.dto';
-export { CreateUserDTO } from './create-user.dto';
