--- src/users/entities/index.ts
@@ -1,1 +1,0 @@
-export { User } from './user.entity';
