--- src/users/index.ts
@@ -1,1 +1,0 @@
-export { UsersModule } from './users.module';
