--- src/users/services/index.ts
@@ -1,1 +1,0 @@
-export { UsersService } from './users.service';
