--- src/utils/abstracts/index.ts
@@ -1,1 +1,0 @@
-export { AbstractEntity } from './abstract-entity';
