+++ src/utils/constants.ts
@@ -1,1 +1,3 @@
-export const passwordRegex = /((?=.*\d)|(?=.*\W+))(?![.\n])(?=.*[A-Z])(?=.*[a-z]).*$/
+export class Constants {
+  static readonly passwordRegex = /((?=.*d)|(?=.*W+))(?![.\n])(?=.*[A-Z])(?=.*[a-z]).*$/;
+}
