--- src/utils/index.ts
@@ -1,2 +1,0 @@
-export { AbstractEntity } from './abstracts';
-export { swaggerConfig } from './swagger-config';
