--- src/utils/mocks/dtos/index.ts
@@ -1,2 +1,0 @@
-export { createUserDTOMock } from './create-user.dto.mock';
-export { createAuthDTOMock } from './create-auth.dto.mock';
