--- src/utils/swagger-config/index.ts
@@ -1,1 +1,0 @@
-export { swaggerConfig } from './swagger-config';
